Move swap logic into classifiers
